chore(storage): move storage/s3-sdk into a proper package
