+++ storage/s3-sdk/src/main/java/storage/s3_sdk/ListGcsObjects.java
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+package storage.s3_sdk;
+
 // [START storage_s3_sdk_list_objects]
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
