+++ storage/s3-sdk/src/test/java/storage/s3_sdk/ListGcsObjectsTest.java
@@ -14,10 +14,13 @@
  * limitations under the License.
  */
 
+package storage.s3_sdk;
+
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 
 import com.amazonaws.services.s3.model.Bucket;
+import com.amazonaws.services.s3.model.ObjectListing;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.util.List;
@@ -27,7 +30,8 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class ListGcsBucketsTest {
+
+public class ListGcsObjectsTest {
   private static final String BUCKET = System.getenv("GOOGLE_CLOUD_PROJECT_S3_SDK");
   private static final String KEY_ID = System.getenv("STORAGE_HMAC_ACCESS_KEY_ID");
   private static final String SECRET_KEY = System.getenv("STORAGE_HMAC_ACCESS_SECRET_KEY");
@@ -60,9 +64,9 @@ public class ListGcsBucketsTest {
   }
 
   @Test
-  public void testListBucket() throws Exception {
-    ListGcsBuckets.listGcsBuckets(KEY_ID, SECRET_KEY);
+  public void testListObjects() throws Exception {
+    ListGcsObjects.listGcsObjects(KEY_ID, SECRET_KEY, BUCKET);
     String output = bout.toString();
-    assertThat(output, CoreMatchers.containsString("Buckets:"));
+    assertThat(output, CoreMatchers.containsString("Objects:"));
   }
 }
